Font width is now measured with floating points.

+++ src/client/java/src/ar/com/tellapic/graphics/IPaintPropertyController.java
@@ -55,7 +55,7 @@ public interface IPaintPropertyController {
 	 * 
 	 * @param size
 	 */
-	public void handleFontSizeChange(int size);
+	public void handleFontSizeChange(float size);
 	
 	/**
 	 *
